fix updateParticipantFirebaseAuth location
