+++ utils/firestore.js
@@ -1888,39 +1888,6 @@ const verifyUsersEmailOrPhone = async (req) => {
     }
 }
 
-const updateParticipantFirebaseAuthentication = async (req, res) => {
-    if(req.method !== 'POST') {
-        return res.status(405).json(getResponseJSON('Only POST requests are accepted!', 405));
-    }
-    const data = req.body.data;
-    const flag = data.flag;
-    const uid =  data.uid;
-
-    if(data === undefined) {
-        return res.status(400).json(getResponseJSON('Bad request. Data is not defined in request body.', 400));
-    }
-
-    let status = '';
-    if (flag === `replaceSignin`) {
-        if (data['phone']) {
-            status = await updateUserPhoneSigninMethod(data.phone, uid);
-        } else if (data['email'] && flag === `replaceSignin`) {
-            status = await updateUserEmailSigninMethod(data.email, uid);
-        }
-    }
-
-    if (flag === `updateEmail` || flag === `updatePhone`) {
-        status = await updateUsersCurrentLogin(data, uid);
-    }
-
-    if (status === true) return res.status(200).json({code: 200});
-    else if (status === `auth/phone-number-already-exists`) return res.status(409).json(getResponseJSON('The user with provided phone number already exists.', 409));
-    else if (status === `auth/email-already-exists`) return res.status(409).json(getResponseJSON('The user with the provided email already exists.', 409));
-    else if (status === `auth/invalid-phone-number`) return res.status(403).json(getResponseJSON('Invalid Phone number', 403));
-    else if (status === `auth/invalid-email`) return res.status(403).json(getResponseJSON('Invalid Email', 403));
-    else return res.status(400).json(getResponseJSON('Operation Unsuccessful', 400));
-}
-
 const updateUsersCurrentLogin = async (req, uid) => {   
     const queries = req
     if (queries.email) {
@@ -2094,5 +2061,4 @@ module.exports = {
     updateUserPhoneSigninMethod,
     updateUserEmailSigninMethod,
     updateUsersCurrentLogin,
-    updateParticipantFirebaseAuthentication
 }
